multiple itcm cores
