+++ src/scriptutil.c
@@ -174,7 +174,7 @@ void code_replacement_apply(CodeReplacement* r, bool apply)
         while ($PC >= r->addr && $PC < r->addr + r->length)
         {
             printf("PC=%x during patch-apply!\n", $PC);
-            __gb_step_cpu(GB);
+            gb_step_cpu(GB);
         }
     }
